+++ src/node.c
@@ -1,17 +1,9 @@
-#include <stdlib.h>
+#include "data.h"
 
 #ifndef NULL
 #define NULL 0
 
 #endif
-
-/* node of the Huffman tree */
-typedef struct node {
-    unsigned int value;
-    char letter;
-    struct node *left;
-    struct node *right;
-} Node;
 
 /* create a node and dynamically allocate memory*/
 Node *create(char c, int freq)
