+++ src/t.c
@@ -1,0 +1,8 @@
+#include "data.h"
+
+int main()
+{
+    *Freq arr[] = {addf('a', 3)}
+    Tree *a = make(arr);
+    print(a->root);
+}
